DL/requirements.txt
# ============================================================================
# Consolidated Requirements for CS59000 Course Project
# ============================================================================
# This file contains all Python dependencies for the entire project.
# Install with: pip install -r requirements.txt
# ============================================================================

# ============================================================================
# Core Web Framework (video-integration backend)
# ============================================================================
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# ============================================================================
# Computer Vision & Media Processing
# ============================================================================
opencv-python==4.8.1.78
mediapipe>=0.10.13

# ============================================================================
# Machine Learning & Deep Learning
# ============================================================================
torch>=2.0.0
transformers>=4.35.0
onnxruntime>=1.16.3
scikit-learn>=1.3.0

# ============================================================================
# Natural Language Processing
# ============================================================================
nltk>=3.8
sentencepiece>=0.1.99

# ============================================================================
# Text-to-Speech
# ============================================================================
piper-tts>=1.0.0

# ============================================================================
# Utilities & Support Libraries
# ============================================================================
numpy>=1.24.3,<2.0.0
requests>=2.31.0
protobuf>=3.20.0,<5.0.0
accelerate>=0.20.0
setuptools>=65.0.0
soundfile>=0.12.0
tqdm>=4.65.0
colorama>=0.4.6


